+++ packages/server/src/constants/environmentVariables.ts
