+++ packages/server/src/db/music_helpers.ts
